+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import RemoteControl from '@/components/features/Common/RemoteControl/RemoteCont
 import Card from '@/components/features/Items/Card/Card';
 import Header from '@/components/widgets/Header/Header';
 import SearchToolbar from '@/components/widgets/SearchToolbar/SearchToolbar';
-import * as styles from '@/styles/rootStyles.css';
+import * as styles from '@/commons/styles/rootStyles.css';
 
 export default function RootPage() {
   return (
